+++ CoolProp/gitrevision.h
@@ -1,1 +1,1 @@
-std::string gitrevision = "9b3ab64937af9f575949e110da5eb06d8e3fa5fa";
+std::string gitrevision = "e21c967ebcc075933163475f1313127672b06933";